Splitting NitriteMapper into two interfaces #74
