+++ nitrite/src/main/java/org/dizitart/no2/tool/JsonDeSerializer.java
@@ -1,0 +1,11 @@
+package org.dizitart.no2.tool;
+
+import org.dizitart.no2.Document;
+
+public interface JsonDeSerializer {
+
+	Document parse(String json);
+
+	String toJson(Object object);
+
+}
